added content to ProductScreen
-List the product details such as image, description, price,
availability etc. for now productScreen component matches the :id parameter
of the url with the product with same id from list of products in
products.js file(later it will be fetched from database)

-Used Row, Col, ListGroup, Card, Button from react-bootstrap for
styling.also a dummy column showing price and add to cart which is not
functional yet.
-also used Rating component inside of ProductScreen which shows the
Rating of the current product

-minor css changes in Footer component